\name{popsom-deprecated}
\alias{popsom-deprecated}
\title{
Deprecated Functions in Package popsom (4.3.0)
}
\description{
These functions have been deprecated as of this release (version 4.3.0) of the popsom package for R. 

Said functions have been flagged for removal upon release of the next version of the popsom package.
}
\details{
The functions currently included in the popsom that have been deprecated are as follows:

1) map.embed() 

-- (NOTE: To explicitly compute the embedding accuracy of maps in future popsom package releases, call map.convergence() with the argument 'verb = True')

2) map.topo() 

-- (NOTE: To explicitly compute the estimated topographic accuracy of maps in future popsom package releases, call map.convergence() with the argument 'verb = True')

3) map.projection()
}
\author{
Lutz Hamel, Benjamin Ott, Greg Breard, University of Rhode Island
               with Robert Tatoian, Vishakh Gopu, and Michael Eiger
}
  }
}
